requirements.txt
python-dotenv==1.0.0
crewai>=0.22.5
crewai-tools>=0.4.26
langchain-openai>=0.1.0
openpyxl>=3.1.0
nicegui>=1.4.0
pandas>=2.0.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
pydantic>=2.5.0
PyYAML>=6.0.0
langfuse>=2.0.0
openinference-instrumentation-crewai>=0.1.0
openinference-instrumentation-litellm>=0.1.0
firecrawl-py>=0.0.8
urllib3>=2.0.0
streamlit>=1.28.0